test(ecology): update plan/apply pipeline for M3 vegetation
